[nextest-runner] move OS-specific code into separate files
More thinning out of `runner/imp.rs`.
